+++ CHANGELOG.md
@@ -5,3 +5,4 @@ FEATURES:
 * **New Resource:** `netapp-gcp_volume`
 * **New Resource:** `netapp-gcp_snapshot`
 * **New Resource:** `netapp-gcp_active_directory`
+* **New Resource:** `netapp-gcp_volume_backup`